LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerlock/pointerevent_coordinates_when_locked-expected.txt

Pointer Events movement in locked state test

Follow the test instructions with mouse. If you don't have the device skip it.

Test Description: This test checks the pointer event coordinates stays unchanged when pointer is locked.
Click left mouse button on the green rectangle.
Move the mouse around.
Click left mouse button again.
Test passes if the proper behavior of the events is observed.


FAIL mouse Test pointerevent coordinates when pointer is locked assert_equals: clientX expected 172 but got 314
